requirements.txt
absl-py==0.13.0
appdirs==1.4.3
apturl==0.5.2
asn1crypto==0.24.0
attrs==21.2.0
audioread==2.1.9
autobahn==17.10.1
Automat==0.8.0
backcall==0.1.0
bcrypt==3.1.7
binwalk==2.2.0
blinker==1.4
bokeh==2.3.2
Brlapi==0.7.0
Brotli==1.0.7
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
cbor==1.0.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
devscripts===2.20.2ubuntu2
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
duplicity==0.8.12.0
EditorConfig==0.12.2
efficientnet-pytorch==0.6.3
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
fonttools==4.5.0
fs==2.4.11
future==0.18.2
gpg===1.13.1-unknown
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.9.0
incremental==16.10.1
intervaltree==3.1.0
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.15.2
Jinja2==3.0.1
joblib==1.0.1
jsbeautifier==1.10.3
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
librosa==0.8.1
llvmlite==0.36.0
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
matplotlib==3.1.2
meson==0.53.2
mido==1.2.10
monotonic==1.5
more-itertools==8.8.0
mpi4py==3.0.3
mpmath==1.1.0
munch==2.5.0
music21==6.7.1
netifaces==0.10.4
nose==1.3.7
note-seq==0.0.3
numba==0.53.1
numpy==1.17.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pandas==1.3.0
paramiko==2.6.0
parso==0.5.2
pefile==2019.4.18
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.3.1
pooch==1.4.0
pretrainedmodels==0.7.4
pretty-midi==0.2.9
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
pydub==0.25.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-gnupg==0.4.5
python-magic==0.4.16
python-snappy==0.5.3
PyTrie==0.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
ranger-fm==1.9.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
resampy==0.2.2
roman==2.0.0
rosdep-modules==0.21.0
rosdistro==0.8.3
rosdistro-modules==0.8.3
rospkg==1.3.0
rospkg-modules==1.3.0
scikit-learn==0.24.2
scipy==1.3.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
segmentation-models-pytorch==0.2.0
selenium==3.141.0
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sympy==1.5.1
system-service==0.3
systemd-python==234
threadpoolctl==2.1.0
timm==0.4.12
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
Twisted==18.9.0
txaio==2.10.0
typing-extensions==3.10.0.0
u-msgpack-python==2.1
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.5
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.1.8
webcolors==1.11.1
wsaccel==0.6.2
xkit==0.0.0
zope.interface==4.7.1
